refactor: change package location 도메인과 wrapper 클래스 분리
