+++ src/main/java/nextstep/ladder/view/ResultView.java
@@ -1,6 +1,10 @@
 package nextstep.ladder.view;
 
 import nextstep.ladder.domain.*;
+import nextstep.ladder.wrapper.Ladder;
+import nextstep.ladder.wrapper.LadderRewards;
+import nextstep.ladder.wrapper.Line;
+import nextstep.ladder.wrapper.Participants;
 
 import java.util.List;
 import java.util.stream.Collectors;
